Initial commit.
Added an app icon. Created the main color schemes for both dark and light modes and the branding related colors.

Added a Tab Bar Controller and filled it with 3 separate tab bar items. One for player stats, one where the actual quiz starts and one for settings.

Ordered the project to follow MVVM.