Added some more stuff the the ResultVC
Now it has a draft collectionView where the quiz questions will be displayed. A score, some labels. The aim is to make the whole collectionView responsive so when the user taps on a cell it displays the details of the question including the source for the correct answer.
I aslo wanna add a footer to it because in the future I wanna add games with more questions or endless mode.